#770 add dependency rule
