+++ wls-common/testing/src/test/java/de/muenchen/oss/wahllokalsystem/wls/common/testing/archunit/utilityClasses/correctFileNamingAndDependencies/service/ThisIsAValidModel.java
@@ -1,0 +1,4 @@
+package de.muenchen.oss.wahllokalsystem.wls.common.testing.archunit.utilityClasses.correctFileNamingAndDependencies.service;
+
+public record ThisIsAValidModel() {
+}
