--- wls-common/testing/src/test/java/de/muenchen/oss/wahllokalsystem/wls/common/testing/archunit/utilityClasses/incorrectFilenaming/domain/ExampleEntity.java
@@ -1,4 +1,0 @@
-package de.muenchen.oss.wahllokalsystem.wls.common.testing.archunit.utilityClasses.incorrectFilenaming.domain;
-
-public class ExampleEntity {
-}
