+++ wls-common/testing/src/test/java/de/muenchen/oss/wahllokalsystem/wls/common/testing/archunit/utilityClasses/incorrectFilenamingAndDependencies/domain/ExampleEntity.java
@@ -1,0 +1,7 @@
+package de.muenchen.oss.wahllokalsystem.wls.common.testing.archunit.utilityClasses.incorrectFilenamingAndDependencies.domain;
+
+import de.muenchen.oss.wahllokalsystem.wls.common.testing.archunit.utilityClasses.incorrectFilenamingAndDependencies.service.DataModelClass;
+
+public class ExampleEntity {
+    DataModelClass dataModelClass = new DataModelClass();
+}
