+++ wls-common/testing/src/test/java/de/muenchen/oss/wahllokalsystem/wls/common/testing/archunit/utilityClasses/incorrectFilenamingAndDependencies/domain/ExampleModel.java
@@ -1,0 +1,4 @@
+package de.muenchen.oss.wahllokalsystem.wls.common.testing.archunit.utilityClasses.incorrectFilenamingAndDependencies.domain;
+
+public class ExampleModel {
+}
