+++ wls-common/testing/src/test/java/de/muenchen/oss/wahllokalsystem/wls/common/testing/archunit/utilityClasses/incorrectFilenamingAndDependencies/rest/DataModelClass.java
@@ -1,0 +1,7 @@
+package de.muenchen.oss.wahllokalsystem.wls.common.testing.archunit.utilityClasses.incorrectFilenamingAndDependencies.rest;
+
+import de.muenchen.oss.wahllokalsystem.wls.common.testing.archunit.utilityClasses.incorrectFilenamingAndDependencies.domain.ExampleDTO;
+
+public record DataModelClass() {
+    static ExampleDTO exampleDTO = new ExampleDTO();
+}
