+++ wls-common/testing/src/test/java/de/muenchen/oss/wahllokalsystem/wls/common/testing/archunit/utilityClasses/incorrectFilenamingAndDependencies/rest/ExampleDto.java
@@ -1,0 +1,7 @@
+package de.muenchen.oss.wahllokalsystem.wls.common.testing.archunit.utilityClasses.incorrectFilenamingAndDependencies.rest;
+
+import de.muenchen.oss.wahllokalsystem.wls.common.testing.archunit.utilityClasses.correctFileNamingAndDependencies.service.ExampleModel;
+
+public record ExampleDto() {
+    static ExampleModel exampleModel = new ExampleModel();
+}
